+++ packages/mention/src/utils/__test__/getSearchTextAt.test.ts
@@ -33,4 +33,24 @@ describe('getSearchTextAt', () => {
     };
     expect(getSearchTextAt('Max', 3, [''])).toEqual(expected);
   });
+
+  it('finds the matching string even if it contains the trigger', () => {
+    const expected = {
+      matchingString: 'one@example.c',
+      begin: 0,
+      end: 14,
+    };
+    expect(getSearchTextAt('@one@example.c', 14, trigger)).toEqual(expected);
+  });
+
+  it('finds the matching string, including trigger, following a trigger', () => {
+    const expected = {
+      matchingString: 'one@example.c',
+      begin: 5,
+      end: 19,
+    };
+    expect(getSearchTextAt('@Max @one@example.c', 19, trigger)).toEqual(
+      expected
+    );
+  });
 });