+++ packages/mention/src/utils/getSearchTextAt.ts
@@ -1,3 +1,5 @@
+import escapeRegExp from 'lodash/escapeRegExp';
+
 export interface SearchTextAtResult {
   begin: number;
   end: number;
@@ -12,20 +14,32 @@ export default function getSearchTextAt(
   triggers: string[]
 ): SearchTextAtResult {
   const str = blockText.substr(0, position);
-  const { begin, index } = triggers
-    .map((trigger, triggerIndex) => ({
-      begin: trigger.length === 0 ? 0 : str.lastIndexOf(trigger),
-      index: triggerIndex,
-    }))
-    .reduce((left, right) => (left.begin >= right.begin ? left : right));
-  const matchingString =
-    triggers[index].length === 0
-      ? str
-      : str.slice(begin + triggers[index].length);
+
+  const triggerPattern = `[${triggers
+    .map((trigger) => escapeRegExp(trigger))
+    .join('')}]`;
+
+  const TRIGGER_REGEX = new RegExp(`(\\s|^)(${triggerPattern})`, 'g');
+
+  const matches = str.matchAll(TRIGGER_REGEX);
+
+  let triggerStartIndex = 0;
+  let valueStartIndex = 0;
+
+  for (const match of matches) {
+    const spaceLen = match[1].length;
+    const matchLen = match[2].length;
+
+    triggerStartIndex = (match.index || 0) + spaceLen;
+    valueStartIndex = triggerStartIndex + matchLen;
+  }
+
+  const matchingString = str.slice(valueStartIndex);
+
   const end = str.length;
 
   return {
-    begin,
+    begin: triggerStartIndex,
     end,
     matchingString,
   };